Restructure mock tables and other changes
