+++ server/src/main/scala/uk/gov/homeoffice/drt/testsystem/crunchsystem/TestPersistentStateActors.scala
@@ -1,7 +1,7 @@
 package uk.gov.homeoffice.drt.testsystem.crunchsystem
 
-import actors.DrtStaticParameters.{expireAfterMillis, startOfTheMonth, time48HoursAgo}
-import actors.{ManifestLookups, ManifestLookupsLike}
+import actors.DrtStaticParameters.expireAfterMillis
+import actors.ManifestLookupsLike
 import actors.persistent.arrivals.CirriumLiveArrivalsActor
 import akka.actor.{ActorRef, ActorSystem, Props}
 import uk.gov.homeoffice.drt.crunchsystem.PersistentStateActors
@@ -12,7 +12,6 @@ case class TestPersistentStateActors(system: ActorSystem,
                                      now: () => SDateLike,
                                      minutesToCrunch: Int,
                                      offsetMinutes: Int,
-                                     maxForecastDays: Int,
                                      manifestLookups: ManifestLookupsLike,
                            ) extends PersistentStateActors {
   override val forecastBaseArrivalsActor: ActorRef =
